requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.0.5
attrs==23.2.0
backcall==0.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dacite==1.8.1
debugpy==1.6.2
decorator==5.1.1
deepali==0.1
deprecation==2.1.0
docker-pycreds==0.4.0
executing==0.8.3
filelock==3.13.1
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.0
gitdb==4.0.11
GitPython==3.1.42
hf-deepali==0.6.2
hydra-core==1.3.2
idna==3.6
imageio==2.34.0
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lightning==2.2.1
lightning-utilities==0.10.1
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.3
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.5.5
networkx==3.2.1
nibabel==5.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==21.3
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
plotly==5.20.0
prompt-toolkit==3.0.30
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.8.2
pytorch-lightning==2.2.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
resolver==0.2.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.4.1.post1
scipy==1.12.0
sentry-sdk==1.42.0
setproctitle==1.3.3
SimpleITK==2.3.1
six==1.16.0
smmap==5.0.1
stack-data==0.3.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
torch==2.2.1
torchmetrics==1.3.2
torchvision==0.17.1
tornado==6.2
tqdm==4.66.2
traitlets==5.14.2
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.4
wcwidth==0.2.5
yarl==1.9.4
